Refactor folder placements
