+++ callbacks/homepage/callbacks.py
@@ -1,5 +1,10 @@
 from collections import namedtuple, OrderedDict
-from pages import lift_over, co_expr, tf_enrich, browse_loci
+import callbacks.homepage.util
+import pages.analysis.lift_over as lift_over
+import pages.analysis.co_expr as co_expr
+import pages.analysis.tf_enrich as tf_enrich
+import pages.analysis.browse_loci as browse_loci
+
 from dash import Input, Output, State, html, ctx
 from dash.exceptions import PreventUpdate
 from .util import *
