+++ pages/navigation/analysis_nav.py
